update compose and manifest functions
