+++ src/releaseFiles/compose/index.ts
@@ -1,0 +1,7 @@
+export { composeDeleteBuildProperties } from "./composeDeleteBuildProperties";
+export { getComposePackageImages } from "./getComposePackageImages";
+export { getComposePath } from "./getComposePath";
+export { parseComposeUpstreamVersion } from "./parseComposeUpstreamVersion";
+export { readCompose } from "./readCompose";
+export { updateComposeImageTags } from "./updateComposeImageTags";
+export { writeCompose } from "./writeCompose";
