+++ src/releaseFiles/manifest/compactManifestIfCore.ts
@@ -2,8 +2,9 @@ import fs from "fs";
 import path from "path";
 import yaml from "js-yaml";
 import { releaseFiles } from "../../params";
-import { readManifest, writeManifest } from "./manifest";
 import { SetupWizard } from "../setupWizard/types";
+import { readManifest } from "./readManifest";
+import { writeManifest } from "./writeManifest";
 
 /**
  * Reads manifest and extra files in `buildDir` compacts them in the manifest
@@ -22,6 +23,8 @@ export function compactManifestIfCore(buildDir: string): void {
 
   writeManifest(manifest, format, { dir: buildDir });
 }
+
+// Utils
 
 function readSetupWizardIfExists(buildDir: string): SetupWizard | null {
   const files = fs.readdirSync(buildDir);
